requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
onnxruntime>=1.17
Pillow>=10.0
opencv-python-headless>=4.8
numpy>=1.24
python-jose[cryptography]>=3.3
requests>=2.31
pytest>=8.0
httpx>=0.27
twilio>=9.0
supabase>=2.10.0
psycopg2-binary>=2.9
